[FLAPI-118] Grouped any() methods
Supports the use of any() methods with a group, as in any(int group).
The first time than an 'any' method is hit, it will trigger any methods
in that group.

Grouping in general has been slightly adjusted to be 'first one wins'.
That is, the first method to be hit will survive while the other groups
will be eliminated. If the method that stays is atMost(1), then it
will go away as well. However if it is any() it will stay. Further, if
two any() methods are in the same group, the first will stay and
the second will go.

Also includes some improvements to AbstractCompiledTest to
allow compiling test cases and running them through the mill.