#197 Pluggable name generators
Adds a new method to the builder `useCustomNameGenerator` which allows
the user to provide their own logic for generated names in the output classes.

Each NameGenerator can customize the method names, and the overall class
names. The class names and method names MUST be unique for the given
keys, so customize with care!

Three implementations are available from the get go.
* DefaultNameGenerator - long names
* CondensedNameGenerator - shorter method names
* HashedNameGenerator - shorter method and class names