#194 generics support for annotated helpers
Adds a dependency to `spring-generics`, and uses it to resolve the generics
of method parameters in introspected helper classes.

With this change, it is also possible to do away with the clunky @BlockChain
generics handling, where the user had to provide the generic type of the
`AtomicReference` class. The Spring tools are able to handle this, and so the
annotation parameter has been removed. (BREAKING CHANGE)